pyfaf: repos: rpm_metadata: Add required parameter for URLs
As the constructor collects all remaining positional arguments as URLs,
we end up receiving a tuple (usually 1-tuple) with a list as the first
element. In places where the URLs are iterated as a list, this will
quickly break things that expect strings and not lists.

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>
(cherry picked from commit 3e112dce1c86dd73ac2525bb1add02786b363614)