generic_table: Add package type to filename of lobs
This is sort of a hack because of DNF and retrace-server.

- DNF checks for file extension (.rpm).
- retrace-server (with UseFafPackages) is not able to install packages from FAF
  because of the missing ".rpm" file extension.

With this change the package type (rpm/deb/tgz) is added at the end of
lob names.

I've alse added checks for backwards compatibility.

New Package lobs should be created with the file extension and the old Package lobs
should stay as they are.

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>